Bug 575149 - Add support to resolve 1st arg type on overload methods
The fix will resolve the first argument type of all matching method
overloads. It also make the relevance of the overload completions higher
so they appear first in the completion list to preserve the existing
behavior. Other completions are strictly provided based on the expected
types. This also update the CompletionContext expected types signatures
so that other completion computers can be benefited.

Change-Id: Id842ff7ad064aabfd9e57c89a94d385d37d8a6dc
Signed-off-by: Gayan Perera <gayanper@gmail.com>